Add files via upload
Renames some of the locations so that they aren't relative to cardinal directions on a map and hopefully makes the rest easier to discern from looking at their names via !missing commands or while using Universal Tracker.

Adds in a new region called `Waterway Overpass` that is the connecting hub for Upper Waterway, Workroom, Worker's Break Room and the Water Injection Chamber. This now acts as the `Downed Zombie` (Video Cassette), the `Before the Waterslide` and the `Rook Panel Outside` (now named Rook Panel) region in the sewers, which should also help clear up a logic bug on Claire's side. 

Adds in a new region called `Central Hub` that is the connecting hub for Treatment Pool Room, Lower Waterway and Bottom Waterway. This now acts as the `Downstairs Location` (now named `On Boxes near Stairs`) and 'End of Left Path' (now named `On Ledge in Water by Trash`) region in the sewers. 

`Keys Panel` now properly has it's `Bottom Waterway` region, which should fix the bug mentioned above where `T-Bar Handle` could've been locked behind itself.

This also randomizes `Dispersal Cartridge - Empty` into the pool, and will have world side changes to accompany it. 

Also some slight fixes to bullet count on a couple weapons for Claire, and gives the `GM 79` a single round for count instead of the usual 0 in vanilla due to instances where the player might receive it as their starting weapon.